web_app/requirements.txt
flask
flask-sqlalchemy
flask-login
google-genai
requests
pymysql

